Fixed nullable properties bug
The issue was relying on the bug/bad documentation of the method ReflectionProperty::hasDefaultValue() which was returning a wrong boolean in case of a promoted property.
Since promoted properties are defined in the constructor, I've resolved this issue by just getting the default value from the ReflectionParameter instead of the ReflectionProperty.